% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\name{clear}
\alias{clear}
\title{Clear the internal state of a Dirichlet Tree}
\usage{
clear(dtree)
}
\arguments{
\item{dtree}{a Dirichlet Tree object.}
}
\description{
Destroys the Tree's internal state to revert it back to the original prior.
}
