% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\docType{class}
\name{dirtree.irv}
\alias{dirtree.irv}
\alias{DirichletTreeIRV}
\title{Dirichlet Tree for IRV}
\usage{
dirtree.irv(nCandidates, scale = 1, dirichlet = F, seed = NULL)
}
\arguments{
\item{nCandidates}{the number of candidates in the IRV election.}

\item{scale}{the prior scale of the distribution.}

\item{dirichlet}{a boolean indicating whether or not the prior is to be
a regular Dirichlet distribution as opposed to a Dirichlet Tree.}

\item{seed}{a string seeding the initial random state of the distribution.}
}
\value{
A Dirichlet Tree representing IRV ballots, as an Rcpp module of class `DirichletTreeIRV`.
}
\description{
A Dirichlet Tree structure for IRV ballot types.
}
\author{
Floyd Everest
}
