% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\docType{class}
\name{dirtree.pirv}
\alias{dirtree.pirv}
\alias{PIRVDirichletTree}
\title{Dirichlet Tree for Partial IRV ballots}
\usage{
dirtree.pirv(
  candidates,
  minDepth = 0,
  maxDepth = length(candidates),
  a0 = 1,
  vd = FALSE
)
}
\arguments{
\item{candidates}{A character vector, with each element (must be unique) representing a single candidate.}

\item{minDepth}{the minimum number of candidates which must be specified for a valid ballot.}

\item{maxDepth}{the maximum number of candidates which can be specified for a valid ballot.}

\item{a0}{the prior parameter for the distribution.}

\item{vd}{a boolean value representing whether or not the prior should reduce to a vanilla Dirichlet distribution.}
}
\value{
A Dirichlet Tree representing partial IRV ballots, as an Rcpp module of class `PIRVDirichletTree`.
}
\description{
\code{dirtree.pirv} is used to create a Dirichlet Tree for modelling partially ordered IRV ballots.
}
\section{References}{

\itemize{
  \item F. Everest, M. Blom, P.B. Stark, P.J. Stuckey, V. Teague and D. Vukcevic. (2022). Auditing Ranked Voting Elections with Dirichlet-Tree Models: First Steps.
}
}

\author{
Floyd Everest
}
\keyword{ballot}
\keyword{dirichlet}
\keyword{election}
\keyword{irv}
\keyword{tree}
