% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\name{draw}
\alias{draw}
\title{Draw ballots from the posterior distribution.}
\usage{
draw(dtree, nBallots)
}
\arguments{
\item{dtree}{a Dirichlet Tree object.}

\item{nBallots}{an integer representing the number of ballots to draw.}
}
\value{
A dataframe with rows corresponding to the drawn ballots.
}
\description{
Draws ballots from one realization of the posterior distribution.
}
