% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\name{evaluate.election}
\alias{evaluate.election}
\title{Evaluates the outcome of an election}
\usage{
evaluate.election(ballots)
}
\arguments{
\item{ballots}{a DataFrame representing the complete set of election ballots>}
}
\value{
an integer representing the index of the victorious candidate.
}
\description{
Given a DataFrame input representing the ballots of an
             election, outputs the winner.
}
