% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_ballots.R
\name{PIRVBallots}
\alias{PIRVBallots}
\title{Construct a set of PIRV ballots.}
\usage{
PIRVBallots(x, ...)
}
\arguments{
\item{x}{A character vector representing a single ballot, or a list of character vectors representing multiple ballots.}
}
\value{
A \code{PIRVBallots} object representing the ballot(s).
}
\description{
\code{PIRVBallots} is used to easily construct a set of PIRV ballots.
}
\examples{
PIRVBallots(LETTERS[1:5])
PIRVBallots(list(LETTERS[1:5], LETTERS[6:1]))
}
