% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballots.R
\name{ranked_ballots}
\alias{ranked_ballots}
\title{Construct a set of ranked ballots.}
\usage{
ranked_ballots(x, candidates = NULL, ...)
}
\arguments{
\item{x}{A character vector representing a single ballot, or a list of
character vectors representing multiple ballots.}

\item{candidates}{A character vector of names corresponding to the candidates running in
the election.}

\item{\\dots}{Additional parameters to pass to \code{ranked_ballots}.}
}
\value{
A \code{ranked_ballots} object representing the ballot(s).
}
\description{
Deprecated in favour of functionality from the `prefio` package.
\code{ranked_ballots} is used to easily construct a set of ranked ballots.
}
\examples{
ranked_ballots(LETTERS[1:5])
ranked_ballots(list(LETTERS[1:5], LETTERS[6:1]))

}
