% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballots.R
\name{read_ballots}
\alias{read_ballots}
\title{Read \code{ranked_ballots} from a file.}
\usage{
read_ballots(file)
}
\arguments{
\item{file}{The name of the file to read from, or a character vector of file lines.}
}
\description{
Deprecated in favour of `prefio` plus PrefLib formats.
Reads a set of partial IRV ballots from a file. The file is expected to
follow the ballot:count standard, with a header describing all participating
candidates.
}
