% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{reset.Rcpp_PIRVDirichletTree}
\alias{reset.Rcpp_PIRVDirichletTree}
\title{Clear the internal state of a PIRV Dirichlet Tree}
\usage{
reset.Rcpp_PIRVDirichletTree(dtree)
}
\arguments{
\item{dtree}{a Dirichlet Tree PIRV object.}
}
\description{
Destroys the Tree's internal state to revert it back to the prior.
}
