% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{reset}
\alias{reset}
\title{Clear the internal state of a PIRV Dirichlet Tree}
\usage{
reset(dtree)
}
\arguments{
\item{dtree}{a PIRV Dirichlet Tree object.}
}
\description{
Destroys the Tree's internal state to revert it back to the prior.
}
