% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Draw election outcomes from the posterior distribution.}
\usage{
sample_posterior(dtree, n_elections, n_ballots, n_winners = 1)
}
\arguments{
\item{dtree}{A \code{dirichlet_tree} object.}

\item{n_elections}{An integer representing the number of elections to generate. A higher
number yields higher precision in the output probabilities.}

\item{n_ballots}{An integer representing the total number of ballots cast in the election.}

\item{n_winners}{The number of candidates elected in each election.}
}
\value{
A numeric vector containing the probabilities for each candidate
being elected.
}
\description{
\code{sample_posterior} draws sets of ballots from independent realizations
of the Dirichlet-Tree posterior, then determines the probability for each
candidate being elected by aggregating the results of the social choice
function. See \insertCite{dtree_evoteid;textual}{dirtree.elections} for
details.
}
\references{
\insertAllCited{}
}
