% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{sample_predictive}
\alias{sample_predictive}
\title{Draw ballots from the posterior predictive distribution.}
\usage{
sample_predictive(dtree, n_ballots)
}
\arguments{
\item{dtree}{A \code{dirichlet_tree} object.}

\item{n_ballots}{An integer representing the number of ballots to draw.}
}
\value{
A \code{prefio::preferences} object containing \code{n_ballots}
ballots drawn from a single realisation of the posterior Dirichlet-tree.
}
\description{
\code{sample_predictive} draws ballots from a multinomial distribution with
probabilities obtained from a single realization of the Dirichlet-tree
posterior on the ranked ballots. See
\insertCite{dtree_evoteid;textual}{elections.dtree} for details.
}
\references{
\insertRef{dtree_eis}{elections.dtree}.

\insertRef{dtree_evoteid}{elections.dtree}.
}
