% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\name{sampleLeafProbability}
\alias{sampleLeafProbability}
\title{Sample a single leaf probability from the distribution, given the specifying ballot.}
\usage{
sampleLeafProbability(dtree, ballot)
}
\arguments{
\item{dtree}{a Dirichlet Tree object.}

\item{ballot}{the ballot which specifies the leaf.}
}
\value{
A probability.
}
\description{
Samples the leaf probability for a single ballot, given the ballot which is represented by the leaf node.
}
