% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{sampleMPP}
\alias{sampleMPP}
\title{Sample marginal posterior probabilities for a ballot.}
\usage{
sampleMPP(dtree, n, ballot)
}
\arguments{
\item{dtree}{A PIRV Dirichlet Tree object.}

\item{n}{The number of samples to draw from the posterior.}

\item{ballot}{The ballot to sample posterior probabilities for.}
}
\value{
A list of NumericVectors with \code{n} probabilities, each corresponding to a probability of observing \code{b} under an independent realisation of the posterior distribution.
}
\description{
Draws marginal probabilities for observing a given ballot under the posterior distribution.
}
