% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{samplePosterior}
\alias{samplePosterior}
\title{Draw election outcomes from the posterior distribution.}
\usage{
samplePosterior(dtree, nElections, nBallots, nWinners = 1)
}
\arguments{
\item{dtree}{A PIRV Dirichlet Tree object.}

\item{nElections}{An integer representing the number of elections to generate. A higher number yields higher precision in the output probabilities.}

\item{nBallots}{An integer representing the number of ballots cast in total for each election.}
}
\value{
A NumericVector containing the probabilities for each candidate being elected.
}
\description{
Draws ballots from a realizations of the Dirichlet Tree posterior, and determines the probability for each candidate being elected by aggregating the results.
}
