% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_ballots.R
\name{social.choice.PIRVBallots}
\alias{social.choice.PIRVBallots}
\title{Computes the outcome of the IRV social choice function.}
\usage{
\method{social.choice}{PIRVBallots}(x, nWinners = 1, ...)
}
\arguments{
\item{x}{The set of ballots for which to compute the outcome of the IRV social choice function.}

\item{nWinners}{The number of candidates to elect.}

\item{\\dots}{Unused.}
}
\description{
Computes the outcome of the IRV social choice function.
}
