% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballots.R
\name{social_choice.ranked_ballots}
\alias{social_choice.ranked_ballots}
\title{Evaluate a social choice function on a set of \code{ranked_ballots}.}
\usage{
\method{social_choice}{ranked_ballots}(x, n_winners = 1, fn = "irv", ...)
}
\arguments{
\item{x}{The set of ballots for which to compute the outcome of the IRV
social choice function.}

\item{n_winners}{The number of candidates to elect.}

\item{fn}{The social choice function to use. \describe{
  \item{\code{"irv"}}{The IRV social choice function.}
}}

\item{\\dots}{Unused.}
}
\description{
\code{social_choice.ranked_ballots} reads a set of ranked ballots and
computes an election outcome. The outcome depends on the chosen
social choice function, along with any its associated parameters.
}
