% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_ballots.R
\name{`[.PIRVBallots`}
\alias{`[.PIRVBallots`}
\alias{[.PIRVBallots}
\title{Access Subsets of Ballots.}
\usage{
\method{[}{PIRVBallots}(x, i = NULL)
}
\arguments{
\item{x}{PIRVBallots.}

\item{i}{The index, or vector of indices corresponding to each extracted ballot.}
}
\description{
Extract subsets of ballots by index.
}
