% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballots.R
\name{`[.ranked_ballots`}
\alias{`[.ranked_ballots`}
\alias{[.ranked_ballots}
\title{Access Subsets of Ballots.}
\usage{
\method{[}{ranked_ballots}(x, i = NULL)
}
\arguments{
\item{x}{Some \code{ranked_ballots}.}

\item{i}{The index, or vector of indices corresponding to each ballot to be extracted.}
}
\description{
Extract subsets of ballots by index.
}
