% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\name{update}
\alias{update}
\alias{update.Rcpp_DirichletTreeIRV}
\title{Update a Dirichlet Tree with ballot data.}
\usage{
\method{update}{Rcpp_DirichletTreeIRV}(dtree, ballots)
}
\arguments{
\item{dtree}{a Dirichlet Tree object.}

\item{ballots}{a dataframe with rows representing the ballots to observe.}
}
\description{
Updates a Dirichlet Tree with observed ballots to obtain a new posterior.
}
