% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballots.R
\name{write_ballots}
\alias{write_ballots}
\title{Write \code{ranked_ballots} to a file.}
\usage{
write_ballots(ballots, filename = "", return_lines = FALSE, suppress = FALSE)
}
\arguments{
\item{ballots}{The \code{ranked_ballots} to write to a file.}

\item{filename}{The name of the file to write to, or \code{""} to write output to stdout.}

\item{return_lines}{A flag which determines whether or not the output should be returned as a
character vector}

\item{suppress}{A flag which, when True, suppresses any output to stdout.}
}
\description{
Deprecated in favour of fucntionality from the `prefio` package.
Writes a set of ballots to a new file. This follows the
ballot:count standard, with a header describing the candidates.
}
\examples{
write_ballots(ranked_ballots(c(LETTERS)), tempfile("test.txt"))
write_ballots(ranked_ballots(c(LETTERS)))

}
