% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\docType{class}
\name{dirtree.pirv}
\alias{dirtree.pirv}
\alias{PIRVDirichletTree}
\title{Dirichlet Tree for Partial IRV ballots}
\usage{
dirtree.pirv(candidates, minDepth, alpha0 = 1)
}
\arguments{
\item{candidates}{A character vector, with each element (must be unique) representing a single candidate.}

\item{minDepth}{the minimum number of candidates to be specified for a valid ballot.}

\item{alpha0}{the prior parameter for the distribution.}
}
\value{
A Dirichlet Tree representing partial IRV ballots, as an Rcpp module of class `PIRVDirichletTree`.
}
\description{
A Dirichlet Tree for modelling partially ordered IRV ballots.
}
\author{
Floyd Everest
}
