% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_ballots.R
\name{read.ballots}
\alias{read.ballots}
\title{Read PIRV ballots from a file.}
\usage{
read.ballots(filename)
}
\arguments{
\item{filename}{The name of the file to read from.}
}
\description{
Reads a set of partial IRV ballots from a file. The file is expected to follow the ballot:count standard, with a header describing all participating candidates.
}
