% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{reset}
\alias{reset}
\title{Clear the internal state of a \code{dirichlet_tree} object.}
\usage{
reset(dtree)
}
\arguments{
\item{dtree}{A \code{dirichlet_tree} object.}
}
\value{
The \code{dirichlet_tree} object.
}
\description{
Destroy the Tree's internal state and revert back to the prior.
}
