% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{samplePosterior}
\alias{samplePosterior}
\title{Draw election outcomes from the posterior distribution.}
\usage{
samplePosterior(dtree, nElections, nBallots, nWinners = 1)
}
\arguments{
\item{dtree}{A PIRV Dirichlet Tree object.}

\item{nElections}{An integer representing the number of elections to generate. A higher number yields higher precision in the output probabilities.}

\item{nBallots}{An integer representing the number of ballots cast in total for each election.}

\item{nWinners}{The number of candidates elected in each election.}
}
\value{
A NumericVector containing the probabilities for each candidate being elected.
}
\description{
\code{samplePosterior} draws sets of ballots from independent realizations of the Dirichlet Tree posterior, then determines the probability for each candidate being elected by aggregating the results of the social choice function.
}
\section{References}{

\itemize{
  \item F. Everest, M. Blom, P.B. Stark, P.J. Stuckey, V. Teague and D. Vukcevic. (2022). Auditing Ranked Voting Elections with Dirichlet-Tree Models: First Steps.
}
}

