% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet-tree-irv.R
\name{samplePosterior}
\alias{samplePosterior}
\title{Sample from the posterior distribution on candidate victory outcomes.}
\usage{
samplePosterior(
  dtree,
  nElections,
  nBallots,
  useObserved = T,
  nBatches = nElections/2
)
}
\arguments{
\item{dtree}{a Dirichlet Tree object.}

\item{nElections}{the number of election outcomes to determine.}

\item{nBallots}{the number of ballots per election.}

\item{useObserved}{a logical indicating whether or not the already observed ballots should
be taken into account to determine outcomes. If false, we resample all ballots.}

\item{nBatches}{a performance metric which determines how many batches the elections should
be split into for tuning multithreading performance.}
}
\value{
A dataframe with rows corresponding to the drawn ballots.
}
\description{
Samples elections from the posterior distribution and determines the outcomes.
}
