% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{samplePredictive}
\alias{samplePredictive}
\title{Draw PIRV ballots from the posterior predictive distribution.}
\usage{
samplePredictive(dtree, nBallots)
}
\arguments{
\item{dtree}{a PIRV Dirichlet Tree object.}

\item{nBallots}{an integer representing the number of ballots to draw.}
}
\value{
A list with each element corresponding to a drawn ballot.
}
\description{
\code{samplePredictive} draws ballots from a multinomial distribution with probabilities obtained from a single realization of the Dirichlet Tree posterior.
}
\section{References}{

\itemize{
  \item F. Everest, M. Blom, P.B. Stark, P.J. Stuckey, V. Teague and D. Vukcevic. (2022). Auditing Ranked Voting Elections with Dirichlet-Tree Models: First Steps.
}
}

