% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_ballots.R
\name{social.choice}
\alias{social.choice}
\title{Computes the outcome of the appropriate social choice function.}
\usage{
social.choice(x, ...)
}
\arguments{
\item{x}{The set of ballots for which to compute the outcome of the social choice function.}

\item{\\dots}{Additional parameters to pass to \code{social.choice}.}
}
\description{
Reads a set of ballots, and computes the outcome of the election. The outcome is described by a vector of winning candidates, along with the elimination order of the losing candidates.
}
