% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballots.R
\name{social_choice}
\alias{social_choice}
\title{Compute the outcome of an election.}
\usage{
social_choice(x, ...)
}
\arguments{
\item{x}{The set of ballots for which to compute the outcome of the social
choice function.}

\item{\\dots}{Additional parameters to pass to \code{social_choice}.}
}
\description{
\code{social_choice} reads a set of ballots, and computes the
outcome of the election. The outcome is described by a vector of winning
candidates, along with the elimination order of the losing candidates.
}
\examples{
social_choice(ranked_ballots(list(LETTERS[1], LETTERS[1], LETTERS[2])))

}
