% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{update}
\alias{update}
\alias{update.PIRVDirichletTree}
\title{Update a Dirichlet Tree with PIRV ballot data.}
\usage{
\method{update}{PIRVDirichletTree}(dtree, ballots)
}
\arguments{
\item{dtree}{a Dirichlet Tree object.}

\item{ballots}{a list of PIRV ballots.}
}
\description{
Updates a Dirichlet Tree with observed ballots to obtain a new posterior.
}
