% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_dtree.R
\name{update}
\alias{update}
\alias{update.Rcpp_PIRVDirichletTree}
\title{Update a Dirichlet Tree with PIRV ballot data.}
\usage{
\method{update}{Rcpp_PIRVDirichletTree}(object, ballots, ...)
}
\arguments{
\item{object}{A PIRV Dirichlet Tree object.}

\item{ballots}{a list of PIRV ballots.}

\item{\\dots}{Unused.}
}
\description{
\code{update} updates a Dirichlet Tree with observations to obtain a posterior distribution.
}
\section{References}{

\itemize{
  \item F. Everest, M. Blom, P.B. Stark, P.J. Stuckey, V. Teague and D. Vukcevic. (2022). Auditing Ranked Voting Elections with Dirichlet-Tree Models: First Steps.
}
}

