% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree.R
\name{update}
\alias{update}
\alias{update.dirichlet_tree}
\title{Update a \code{dirichlet_tree} model by observing some ranked ballots.}
\usage{
\method{update}{dirichlet_tree}(object, ballots, ...)
}
\arguments{
\item{object}{A PIRV Dirichlet Tree object.}

\item{ballots}{a list of PIRV ballots.}

\item{\\dots}{Unused.}
}
\value{
The \code{dirichlet_tree} object.
}
\description{
\code{update} updates a Dirichlet-Tree model with observations to obtain
a posterior distribution on the ranked ballots. See
\insertCite{dtree_evoteid;textual}{dirtree.elections} for implementation
details.
}
\references{
\insertAllCited{}
}
