% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirv_ballots.R
\name{write.ballots}
\alias{write.ballots}
\title{Write PIRV ballots to a file.}
\usage{
write.ballots(filename, ballots, candidates = NULL)
}
\arguments{
\item{filename}{The name of the file to write to.}

\item{ballots}{The IRVBallots to write to a file.}

\item{candidates}{Optionally, supply the candidate names in advance.}
}
\description{
Writes a set of partial IRV ballots to a new file. This follows the ballot:count standard, with a header describing candidates.
}
